requirements.txt
streamlit>=1.28.0
boto3>=1.34.0
strands-agents>=1.14.0
uuid